Split mixed pipeline/non-pipeline test files
Three files moved wholesale in the previous commit actually contain a mix of
pipeline-dependent and pipeline-free tests. Split each so every test lives on the
correct side of the tests/integration vs tests/pipeline boundary:

- scripts/console_scripts_test.py: test_pudl_datastore, test_resource_description, and
  test_zenodo_data_release (none touch prebuilt_outputs) move to
  tests/integration/scripts/; only test_pudl_service_territories
  (usefixtures("prebuilt_outputs")) stays in tests/pipeline/scripts/.
- extract/ferc_dbf_extract_test.py: test_ferc_schema (only needs zenodo_datastore,
  self-skips on GHA for disk space) moves to tests/integration/extract/;
  test_ferc1_dbf2sqlite (needs ferc1_engine_dbf) stays in tests/pipeline/extract/.
- validate/dbt_test.py: test_validate_asset_selection (dry-run, no prebuilt outputs
  needed) moves to tests/integration/validate/; test_update_tables
  (usefixtures("prebuilt_outputs")) stays in tests/pipeline/validate/.

Also remove tests/integration/dagster/ and tests/integration/glue/, which contained only
__init__.py after their sole test files (geoparquet_test.py, pipeline_test.py,
glue_test.py) moved to tests/pipeline/ in the prior commit — none of their tests are
pipeline-free.

Verified via `pytest --collect-only` that all 468 tests under tests/integration and
tests/pipeline are still collected with no duplicates or omissions, and `ruff
check`/`ruff format` pass on all changed files.
