+++ tests/pipeline/scripts/console_scripts_test.py
@@ -1,31 +1,9 @@
 """Test the PUDL console scripts from within PyTest."""
 
-import os
 from pathlib import Path
 
 import geopandas as gpd  # noqa: ICN002
 import pytest
-
-
-@pytest.mark.parametrize(
-    "command",
-    [
-        # Force the download of one small partition from Zenodo
-        "pudl_datastore ferc60 --bypass-local-cache --partition year=2020",
-        # Force the download of one small partition from GCS
-        "pudl_datastore ferc60 --bypass-local-cache --cloud-cache-path s3://pudl.catalyst.coop/zenodo --partition year=2020",
-        # Exercise the datastore validation code
-        "pudl_datastore ferc60 --validate --partition year=2020",
-        # Ensure that all data source partitions are legible
-        "pudl_datastore ferc60 eia860 --list-partitions --cloud-cache-path s3://pudl.catalyst.coop/zenodo",
-    ],
-)
-@pytest.mark.script_launch_mode("inprocess")
-def test_pudl_datastore(script_runner, command: str):
-    """CLI tests specific to the pudl_datastore script."""
-    runner_args = command.split(" ")
-    ret = script_runner.run(runner_args, print_result=True)
-    assert ret.success
 
 
 @pytest.mark.parametrize(
@@ -96,61 +74,3 @@ def test_pudl_service_territories(
     gdf = gpd.read_parquet(out_path)
     assert set(gdf.columns) == expected_cols
     assert not gdf.empty
-
-
-@pytest.mark.parametrize(
-    "resource_id,expected_success",
-    [
-        # One migrated table
-        ("core_epacems__hourly_emissions", True),
-        # One unmigrated table (TODO: drop after migration)
-        ("_core_eia860__cooling_equipment", True),
-        # One nonexistent table
-        ("imaginary_resource", False),
-    ],
-)
-@pytest.mark.script_launch_mode("inprocess")
-@pytest.mark.order(1)
-def test_resource_description(script_runner, resource_id: str, expected_success: bool):
-    """CLI tests specific to the resource_description script."""
-    ret = script_runner.run(
-        ["resource_description", "-n", resource_id], print_result=True
-    )
-    assert ret.success is expected_success
-    if not expected_success:
-        assert f"No table {resource_id}" in ret.stdout
-
-
-@pytest.mark.script_launch_mode("inprocess")
-@pytest.mark.skipif(
-    "ZENODO_SANDBOX_TOKEN_PUBLISH" not in os.environ,
-    reason="Zenodo sandbox token required to exercise uploads",
-)
-@pytest.mark.xfail(reason="Sadly the Zenodo sandbox has become flaketastic.")
-def test_zenodo_data_release(script_runner, tmp_path: Path):
-    """Round-trip upload a tiny release directory to the Zenodo sandbox.
-
-    The CLI accepts any fsspec-compatible path. Using ``Path.as_uri()`` yields a
-    ``file://`` prefix so the integration test exercises that codepath while
-    keeping all artifacts confined to pytest's temporary directory management.
-    """
-
-    source_dir = tmp_path / "release"
-    source_dir.mkdir()
-    (source_dir / "README.txt").write_text("Test release contents", encoding="utf-8")
-    (source_dir / "data.csv").write_text("value\n1\n", encoding="utf-8")
-    (source_dir / "metadata.json").write_text("{}\n", encoding="utf-8")
-
-    result = script_runner.run(
-        [
-            "zenodo_data_release",
-            "--env",
-            "sandbox",
-            "--source-dir",
-            source_dir.as_uri(),
-            "--publish",
-        ],
-        print_result=True,
-    )
-
-    assert result.success
